Leave the service collection open to addition until the handler is about to be created, then construct the SP
Also move the CommandSupport from shared into its own library - shared is getting too crowded
