+++ Aspirate.sln
@@ -45,6 +45,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "actions", "actions", "{5644
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Aspirate.Secrets", "src\Aspirate.Secrets\Aspirate.Secrets.csproj", "{5801EDCE-B1D1-474C-8F2B-BB2927372867}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Aspirate.CommandSupport", "src\Aspirate.CommandSupport\Aspirate.CommandSupport.csproj", "{BF110CDB-51B7-4320-84D1-464691D83A88}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -67,6 +69,10 @@ Global
 		{5801EDCE-B1D1-474C-8F2B-BB2927372867}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5801EDCE-B1D1-474C-8F2B-BB2927372867}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5801EDCE-B1D1-474C-8F2B-BB2927372867}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BF110CDB-51B7-4320-84D1-464691D83A88}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BF110CDB-51B7-4320-84D1-464691D83A88}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BF110CDB-51B7-4320-84D1-464691D83A88}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BF110CDB-51B7-4320-84D1-464691D83A88}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -79,6 +85,7 @@ Global
 		{8F2F0EB2-5569-496F-8258-D19D570FC464} = {8B1FEC50-8212-47F4-A2AC-7940C076B1DC}
 		{5644B29D-6060-409C-B1A8-04D7A26EF958} = {8B1FEC50-8212-47F4-A2AC-7940C076B1DC}
 		{5801EDCE-B1D1-474C-8F2B-BB2927372867} = {796B4241-81DD-4233-8DB5-FAE145D55F46}
+		{BF110CDB-51B7-4320-84D1-464691D83A88} = {796B4241-81DD-4233-8DB5-FAE145D55F46}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {4990E3C8-4D86-44BB-BC2B-7D055DBC7958}
