+++ src/Aspirate.Cli/GlobalUsings.cs
@@ -3,17 +3,14 @@
 global using System.CommandLine;
 global using System.CommandLine.Builder;
 global using System.CommandLine.Invocation;
-global using System.CommandLine.NamingConventionBinder;
 global using System.CommandLine.Parsing;
 global using System.IO.Abstractions;
-global using System.Reflection;
 global using System.Text;
 global using System.Text.Json;
 global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.Extensions.DependencyInjection.Extensions;
 
 global using Aspirate.Cli;
-global using Aspirate.Cli.Actions;
 global using Aspirate.Cli.Actions.Configuration;
 global using Aspirate.Cli.Actions.Containers;
 global using Aspirate.Cli.Actions.Manifests;
@@ -31,8 +28,8 @@ global using Aspirate.Cli.Processors.Project;
 global using Aspirate.Cli.Processors.RabbitMQ;
 global using Aspirate.Cli.Processors.Redis;
 global using Aspirate.Cli.Services;
-global using Aspirate.Shared.Actions;
-global using Aspirate.Shared.Commands;
+global using Aspirate.CommandSupport.Actions;
+global using Aspirate.CommandSupport.Commands;
 global using Aspirate.Shared.Exceptions;
 global using Aspirate.Shared.Extensions;
 global using Aspirate.Shared.Literals;
@@ -47,5 +44,4 @@ global using Aspirate.Shared.Processors;
 global using Aspirate.Shared.Services;
 global using CliWrap;
 global using CliWrap.Buffered;
-global using CliWrap.EventStream;
 global using Spectre.Console;
