+++ src/Aspirate.CommandSupport/Aspirate.CommandSupport.csproj
