+++ src/Aspirate.CommandSupport/Commands/BaseCommand.cs
@@ -1,4 +1,4 @@
-namespace Aspirate.Shared.Commands;
+namespace Aspirate.CommandSupport.Commands;
 
 [ExcludeFromCodeCoverage]
 public abstract class BaseCommand<TOptions, TOptionsHandler> : Command
@@ -8,13 +8,13 @@ public abstract class BaseCommand<TOptions, TOptionsHandler> : Command
     protected BaseCommand(string name, string description)
         : base(name, description)
     {
-        Handler = CommandHandler.Create<TOptions, IServiceProvider>(HandleOptions);
+        Handler = CommandHandler.Create<TOptions, IServiceCollection>(ConstructCommand);
         AddOption(NonInteractive);
     }
 
-    private static Task<int> HandleOptions(TOptions options, IServiceProvider serviceProvider)
+    private static Task<int> ConstructCommand(TOptions options, IServiceCollection services)
     {
-        var handler = ActivatorUtilities.CreateInstance<TOptionsHandler>(serviceProvider);
+        var handler = ActivatorUtilities.CreateInstance<TOptionsHandler>(services.BuildServiceProvider());
 
         handler.CurrentState.PopulateStateFromOptions(options);
