+++ src/Aspirate.CommandSupport/Commands/ICommandOptions.cs
@@ -1,0 +1,3 @@
+namespace Aspirate.CommandSupport.Commands;
+
+public interface ICommandOptions;
