--- src/Aspirate.Shared/Commands/ICommandOptions.cs
@@ -1,3 +1,0 @@
-namespace Aspirate.Shared.Commands;
-
-public interface ICommandOptions;
