+++ tests/Aspirate.Tests/GlobalUsings.cs
@@ -7,13 +7,11 @@ global using Aspirate.Cli.Actions.Containers;
 global using Aspirate.Cli.Actions.Manifests;
 global using Aspirate.Cli.Extensions;
 global using Aspirate.Cli.Services;
+global using Aspirate.CommandSupport.Actions;
 global using Aspirate.Secrets.Providers.Base64;
 global using Aspirate.Secrets.Providers.Password;
-global using Aspirate.Shared.Actions;
 global using Aspirate.Shared.Exceptions;
 global using Aspirate.Shared.Extensions;
-global using Aspirate.Shared.Literals;
-global using Aspirate.Shared.Models;
 global using Aspirate.Shared.Models.Aspirate;
 global using Aspirate.Shared.Models.AspireManifests.Components;
 global using Aspirate.Shared.Models.AspireManifests.Components.V0;
